move UI config to test folder and allow optionally to define a local config that will not be commited
