+++ tests/UI/config.dist.js
